fix(slack): split long agent notifications instead of truncating them
An agent posting through the notify route had its message cut at 2900
characters, mid-sentence, with no way to recover it. The reply reported
truncated: true, so the agent knew — after the fact, with the text already gone.
In practice the cut lands on the tail, which is where a writeup keeps its
conclusions and recommendations.

The completion path already solved this: splitIntoSlackSections spreads long
text across consecutive section blocks, keeping code fences balanced across the
seam. The notify route never got the same treatment, so the same worker had two
different answers to the same Slack limit.

Move the splitter into @open-inspect/shared/slack and use it in both places. The
notify route now sanitizes against the raw-input cap it already enforces, splits
the result, and emits one section per chunk; SLACK_TEXT_MAX_LENGTH is gone. The
splitter's own ceiling sits far above that input cap, so truncated now only
reports the raw-input limit rather than a Slack rendering detail.

The splitter's tests deliberately stay in the slack-bot completion suite: they
exercise it through buildCompletionBlocks and so also assert that moving it left
the completion output unchanged.
