+++ packages/control-plane/src/routes/slack-notify.ts
@@ -7,6 +7,7 @@ import {
   getPermalink,
   postMessage,
   sanitizeAgentText,
+  splitIntoSlackSections,
   SLACK_DENIAL_STATUS,
   type SlackNotifySuccessOutput,
   type SlackWireDenialReason,
@@ -20,9 +21,12 @@ import { error, json, type RequestContext } from "./shared";
 
 const logger = createLogger("slack-notify");
 
-/** Maximum text length before truncation; fits within Slack's section block. */
-const SLACK_TEXT_MAX_LENGTH = 2900;
-/** Hard cap on the raw text we accept and persist verbatim in event args. */
+/**
+ * Hard cap on the raw text we accept and persist verbatim in event args. Also
+ * the sanitizer's ceiling: text longer than one Slack section is split across
+ * consecutive sections rather than cut, so the section limit is not a limit on
+ * what an agent may post.
+ */
 const RAW_TEXT_INPUT_MAX_LENGTH = 12_000;
 /** Channel name length cap (Slack max is 80). */
 const CHANNEL_INPUT_MAX_LENGTH = 80;
@@ -103,7 +107,7 @@ export async function handleSlackNotify(
 
   const sanitized = sanitizeAgentText(parsed.text, {
     mentionsPolicy,
-    maxLength: SLACK_TEXT_MAX_LENGTH,
+    maxLength: RAW_TEXT_INPUT_MAX_LENGTH,
   });
 
   if (sanitized.text.trim().length === 0) {
@@ -114,14 +118,16 @@ export async function handleSlackNotify(
     );
   }
 
+  const sections = splitIntoSlackSections(sanitized.text);
   const blocks = buildBlocks({
-    text: sanitized.text,
+    sections,
     sessionId,
     appName: env.APP_NAME ?? "Open-Inspect",
     webAppUrl: env.WEB_APP_URL,
   });
-
-  const post = await postMessage(token, parsed.channel, sanitized.text, {
+  // A section-split message has no single text body; the notification preview
+  // takes the opening section.
+  const post = await postMessage(token, parsed.channel, sections[0] ?? sanitized.text, {
     thread_ts: parsed.threadTs,
     blocks,
   });
@@ -143,6 +149,9 @@ export async function handleSlackNotify(
     channelId,
     messageTs,
     permalink,
+    // Only the raw-input cap can truncate now: the splitter's own ceiling
+    // (MAX_RESPONSE_SECTIONS sections) is far above RAW_TEXT_INPUT_MAX_LENGTH,
+    // and text that merely exceeds one section is split rather than cut.
     truncated: sanitized.truncated,
     strippedBroadcasts: sanitized.strippedBroadcasts,
     mentionsModified: sanitized.mentionsModified,
@@ -211,16 +220,16 @@ async function parseBody(request: Request): Promise<ParsedBody | Response> {
 }
 
 function buildBlocks(opts: {
-  text: string;
+  sections: string[];
   sessionId: string;
   appName: string;
   webAppUrl: string | undefined;
 }): unknown[] {
   const blocks: unknown[] = [
-    {
+    ...opts.sections.map((section) => ({
       type: "section",
-      text: { type: "mrkdwn", text: opts.text },
-    },
+      text: { type: "mrkdwn", text: section },
+    })),
     {
       type: "context",
       elements: [
